Struct iobuf::AROIobuf

pub struct AROIobuf {
    // some fields omitted
}

Atomic Read-Only Iobuf

An ROIobuf which is safe to Send across tasks and Share with other tasks. This atomically refcounts the buffer, which has a greater cost on .clone(), but provides more flexibility to multithreaded consumers.

To create an AROIobuf, create a normal Iobuf and call .atomic_read_only().

Below is an example of fill an Iobuf in one thread with the numbers 0x00 to 0xFF, and consuming/validating these numbers in parallel in 4 other threads:

#![feature(std_misc)]
use iobuf::{RWIobuf, AROIobuf, Iobuf};
use std::thread;

// Write the bytes 0x00 - 0xFF into an Iobuf.
fn fill(buf: &mut RWIobuf<'static>) -> Result<(), ()> {
  for i in 0x00u32 .. 0x100 {
    try!(buf.fill_be(i as u8));
  }

  Ok(())
}

// Validates the contents of buf are `idx`, `idx+1`, `idx+2`, ...
// until the buffer is exhausted.
fn check(buf: &mut AROIobuf, mut idx: u32) -> Result<(), ()> {
  loop {
    let b: u8 =
      match buf.consume_be::<u8>() {
        Err(()) => return Ok(()),
        Ok(b) => b
      };

    if b as u32 == idx {
      idx += 1;
    } else {
      return Err(())
    }
  }
}

let mut source_b = RWIobuf::new(256);
assert_eq!(fill(&mut source_b), Ok(()));

// Reset the Iobuf for reading.
source_b.flip_lo();

// We can still clone this buffer. It will be non-atomically refcounted for
// now.
{
  let _ = source_b.clone();
}

// Now prepare it for sending to our 4 subtasks.
let shared_b: AROIobuf = source_b.atomic_read_only().ok().unwrap();

let mut tasks = vec!();

for i in 0..4 {
  // This clone modifies the AROIobuf's atomic refcount.
  let mut b = shared_b.clone();
  tasks.push(thread::spawn(move || {
    let start = i*0x40;
    assert_eq!(b.advance(start), Ok(()));
    assert_eq!(b.resize(0x40), Ok(()));
    assert_eq!(check(&mut b, start), Ok(()));
    assert!(b.is_empty());

    // This clone will atomically modify refcounts.
    {
      let _ = b.clone();
    }
  }));
}

for mut t in tasks.into_iter() {
  t.join();
}

Methods

impl AROIobuf

fn read_only(self) -> Result<ROIobuf<'static>, AROIobuf>

Stops atomically reference counting a unique buffer. This method returns Ok if the AROIobuf is the last of its kind, and Err if it's not.

use iobuf::{AROIobuf, ROIobuf, Iobuf};

let buf: ROIobuf<'static> = ROIobuf::from_str_copy("hello, world!");
let a_buf: AROIobuf = buf.atomic_read_only().ok().unwrap();
unsafe { assert_eq!(a_buf.as_window_slice(), b"hello, world!"); }

let buf: ROIobuf<'static> = a_buf.read_only().ok().unwrap();

// TA-DA!
unsafe { assert_eq!(buf.as_window_slice(), b"hello, world!"); }

fn read_write(self) -> Result<RWIobuf<'static>, AROIobuf>

Stops atomically reference counting a unique buffer. This method returns Ok if the AROIobuf is the last of its kind, and Err if it's not.

use iobuf::{AROIobuf, ROIobuf, RWIobuf, Iobuf};

let buf: ROIobuf<'static> = ROIobuf::from_str_copy("hello, world!");
let a_buf: AROIobuf = buf.atomic_read_only().ok().unwrap();
unsafe { assert_eq!(a_buf.as_window_slice(), b"hello, world!"); }

let buf: RWIobuf<'static> = a_buf.read_write().ok().unwrap();

// TA-DA!
unsafe { assert_eq!(buf.as_window_slice(), b"hello, world!"); }