Struct iobuf::RWIobuf

pub struct RWIobuf<'a> {
    // some fields omitted
}

Read-Write Iobuf

An Iobuf which can read and write into a buffer.

Iobufs may be cloned cheaply. When cloned, the data itself is shared and refcounted, and a new copy of the limits and window is made. This can be used to construct multiple views into the same buffer.

poke and fill write a value at a position relative to the start of the window. Only fill advances the window by the amount written. They are meant to be used with try!.

A suffix _be means the data will be read big-endian. A suffix _le means the data will be read little-endian.

The unsafe_ prefix means the function omits bounds checks. Misuse can easily cause security issues. Be careful!

Methods

impl<'a> RWIobuf<'a>

fn empty() -> RWIobuf<'static>

Constructs a trivially empty Iobuf, limits and window are 0, and there's an empty backing buffer. This will not allocate.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::empty();

assert_eq!(b.len(), 0);
assert_eq!(b.cap(), 0);

fn new(len: usize) -> RWIobuf<'static>

Constructs a new Iobuf with a buffer of size len, undefined contents, and the limits and window set to the full size of the buffer.

The maximum length of an Iobuf is approximately 2 GB.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::new(10);

assert_eq!(b.len(), 10);
assert_eq!(b.cap(), 10);

fn new_with_allocator(len: usize, allocator: Arc<Box<Allocator>>) -> RWIobuf<'static>

Constructs a new Iobuf with a buffer of size len, undefined contents, and the limits and window set to the full range of the buffer. The memory will be allocated out of the given allocator, instead of the global heap.

The maximum length of an Iobuf is approximately 2 GB.

fn from_str_copy(s: &str) -> RWIobuf<'static>

Copies a str into a writeable Iobuf. The contents of the str will be copied, so prefer to use the non-copying constructors whenever possible.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::from_str_copy("hello");

assert_eq!(b.poke_be(1, b'4'), Ok(()));

assert_eq!(b.len(), 5);
assert_eq!(b.cap(), 5);
unsafe { assert_eq!(b.as_window_slice(), b"h4llo"); }
unsafe { assert_eq!(b.as_limit_slice(), b"h4llo"); }

fn from_str_copy_with_allocator(s: &str, allocator: Arc<Box<Allocator>>) -> RWIobuf<'static>

Copies a str into a writeable Iobuf, whose memory comes from the given allocator.

fn from_slice(s: &'a mut [u8]) -> Self

Constructs an Iobuf from a slice. The Iobuf will not copy the slice contents, and therefore their lifetimes will be linked.

use iobuf::{RWIobuf,Iobuf};

let mut s = [1,2,3,4];

{
  let mut b = RWIobuf::from_slice(&mut s[..]);

  assert_eq!(b.advance(1), Ok(()));
  assert_eq!(b.peek_be(1), Ok(0x0304u16)); // ...and the Iobuf!
  assert_eq!(b.poke_be(1, 100u8), Ok(()));
}

// We can still use the slice, but only because of the braces above.
assert_eq!(s[2], 100);

fn from_slice_copy(s: &[u8]) -> RWIobuf<'static>

Copies a byte vector into a new, writeable Iobuf. The contents of the slice will be copied, so prefer to use the other constructors whenever possible.

use iobuf::{RWIobuf,Iobuf};

let mut v = vec!(1u8, 2, 3, 4, 5, 6, 10);
v[1] = 20;

let mut b = RWIobuf::from_slice_copy(&v[..]);

let expected = [ 1,20,3,4,5,6,10 ];
unsafe { assert_eq!(b.as_window_slice(), expected); }

fn from_slice_copy_with_allocator(s: &[u8], allocator: Arc<Box<Allocator>>) -> RWIobuf<'static>

Copies a byte vector into a new writeable Iobuf, whose memory comes from the given allocator.

unsafe fn as_mut_window_slice<'b>(&'b self) -> &'b mut [u8]

Reads the data in the window as a mutable slice. Note that since &mut in rust really means &unique, this function lies. There can exist multiple slices of the same data. Therefore, this function is unsafe.

It may only be used safely if you ensure that the data in the iobuf never interacts with the slice, as they point to the same data. peeking or pokeing the slice returned from this function is a big no-no.

use iobuf::{RWIobuf, Iobuf};

let mut s = [1,2,3];

{
  let mut b = RWIobuf::from_slice(&mut s[..]);

  assert_eq!(b.advance(1), Ok(()));
  unsafe { b.as_mut_window_slice()[1] = 30; }
}

let expected = [ 1,2,30 ];
assert_eq!(s, &expected[..]);

unsafe fn as_mut_limit_slice<'b>(&'b self) -> &'b mut [u8]

Reads the data in the window as a mutable slice. Note that since &mut in rust really means &unique, this function lies. There can exist multiple slices of the same data. Therefore, this function is unsafe.

It may only be used safely if you ensure that the data in the iobuf never interacts with the slice, as they point to the same data. peeking or pokeing the slice returned from this function is a big no-no.

use iobuf::{RWIobuf, Iobuf};

let mut s = [1,2,3];

{
  let mut b = RWIobuf::from_slice(&mut s[..]);

  assert_eq!(b.advance(1), Ok(()));
  unsafe { b.as_mut_limit_slice()[1] = 20; }
}

assert_eq!(s, &[1,20,3][..]);

fn read_only(&self) -> ROIobuf<'a>

Gets a read-only copy of this Iobuf. This is a very cheap operation, as the backing buffers are shared. This can be useful for interfacing with code that only accepts read-only Iobufs.

In general, ROIobuf should never be used as a function parameter. If read-only acceess is all that is required, take a generic <T: Iobuf>.

use iobuf::{ROIobuf,RWIobuf,Iobuf};

let mut s = [1,2,3,4];

let rwb: RWIobuf = RWIobuf::from_slice(&mut s[..]);

// write some data into rwb.

let rb: ROIobuf = rwb.read_only();

// now do read-only ops.
assert_eq!(rb.len(), 4);

fn compact(&mut self)

Copies data from the window to the lower limit fo the iobuf and sets the window to range from the end of the copied data to the upper limit. This is typically called after a series of Consumes to save unread data and prepare for the next series of Fills and flip_los.

use std::result::Result::{self,Ok};
use iobuf::{RWIobuf,Iobuf};

// A header, saying how many shorts will follow. Unfortunately, our buffer
// isn't big enough for all the shorts! Assume the rest will be sent in a
// later packet.
let mut s = [ 0x02, 0x11, 0x22, 0x33 ];
let mut b = RWIobuf::from_slice(&mut s[..]);

#[derive(Eq, PartialEq, Debug)]
enum ParseState {
  NeedMore(u16), // sum so far
  Done(u16),     // final sum
};

// Returns a pair of the sum of shorts seen so far, and `true` if we're
// finally done parsing. The sum will be partial if parsing is incomplete.
fn parse(b: &mut RWIobuf) -> Result<ParseState, ()> {
  let len: u8 = try!(b.consume_be());
  let mut sum = 0u16;

  for _ in (0u8 .. len) {
    unsafe {
      if b.len() < 2 {
        b.compact();
        return Ok(ParseState::NeedMore(sum));
      }
      sum += b.unsafe_consume_be::<u16>();
    }
  }

  Ok(ParseState::Done(sum))
}

assert_eq!(parse(&mut b), Ok(ParseState::NeedMore(0x1122)));
assert_eq!(b.len(), 3);
assert_eq!(b.cap(), 4);
assert_eq!(b.peek_be(0), Ok(0x11u8));

fn poke(&self, pos: u32, src: &[u8]) -> Result<(), ()>

Writes the bytes at a given offset from the beginning of the window, into the supplied buffer. Either the entire buffer is copied, or an error is returned because bytes outside of the window would be written.

use iobuf::{RWIobuf,Iobuf};

let data = [ 1,2,3,4 ];

let mut b = RWIobuf::new(10);

assert_eq!(b.poke(0, &data[..]), Ok(()));
assert_eq!(b.poke(3, &data[..]), Ok(()));
assert_eq!(b.resize(7), Ok(()));
assert_eq!(b.poke(4, &data[..]), Err(())); // no partial write, just failure

let expected = [ 1,2,3,1,2,3,4 ];
unsafe { assert_eq!(b.as_window_slice(), expected); }

fn poke_be<T: IntLike>(&self, pos: u32, t: T) -> Result<(), ()>

Writes a big-endian primitive at a given offset from the beginning of the window.

An error is returned if bytes outside of the window would be accessed.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::new(10);

assert_eq!(b.poke_be(0, 0x0304u16), Ok(()));
assert_eq!(b.poke_be(1, 0x0505u16), Ok(()));
assert_eq!(b.poke_be(3, 0x06070809u32), Ok(()));

assert_eq!(b.resize(7), Ok(()));

let expected = [ 3,5,5,6,7,8,9 ];
unsafe { assert_eq!(b.as_window_slice(), expected); }

fn poke_le<T: IntLike>(&self, pos: u32, t: T) -> Result<(), ()>

Writes a little-endian primitive at a given offset from the beginning of the window.

An error is returned if bytes outside of the window would be accessed.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::new(10);

assert_eq!(b.poke_le(0, 0x0304u16), Ok(()));
assert_eq!(b.poke_le(1, 0x0505u16), Ok(()));
assert_eq!(b.poke_le(3, 0x06070809u32), Ok(()));

assert_eq!(b.resize(7), Ok(()));

unsafe { assert_eq!(b.as_window_slice(), [ 4, 5, 5, 9, 8, 7, 6 ]); }

fn fill(&mut self, src: &[u8]) -> Result<(), ()>

Writes bytes from the supplied buffer, starting from the front of the window. Either the entire buffer is copied, or an error is returned because bytes outside the window were requested.

After the bytes have been written, the window will be moved to no longer include then.

use iobuf::{RWIobuf,Iobuf};

let data = [ 1, 2, 3, 4 ];

let mut b = RWIobuf::new(10);

assert_eq!(b.fill(&data[..]), Ok(()));
assert_eq!(b.fill(&data[..]), Ok(()));
assert_eq!(b.fill(&data[..]), Err(()));

b.flip_lo();

unsafe { assert_eq!(b.as_window_slice(), &[ 1,2,3,4,1,2,3,4 ][..]); }

fn fill_be<T: IntLike>(&mut self, t: T) -> Result<(), ()>

Writes a big-endian primitive into the beginning of the window.

After the primitive has been written, the window will be moved such that it is no longer included.

An error is returned if bytes outside of the window were requested.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::new(10);

assert_eq!(b.fill_be(0x12345678u32), Ok(()));
assert_eq!(b.fill_be(0x11223344u32), Ok(()));
assert_eq!(b.fill_be(0x54321123u32), Err(()));
assert_eq!(b.fill_be(0x8877u16), Ok(()));

b.flip_lo();

unsafe { assert_eq!(b.as_window_slice(), [ 0x12, 0x34, 0x56, 0x78
                                         , 0x11, 0x22, 0x33, 0x44
                                         , 0x88, 0x77 ]); }

fn fill_le<T: IntLike>(&mut self, t: T) -> Result<(), ()>

Writes a little-endian primitive into the beginning of the window.

After the primitive has been written, the window will be moved such that it is no longer included.

An error is returned if bytes outside of the window were requested.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::new(10);

assert_eq!(b.fill_le(0x12345678u32), Ok(()));
assert_eq!(b.fill_le(0x11223344u32), Ok(()));
assert_eq!(b.fill_le(0x54321123u32), Err(()));
assert_eq!(b.fill_le(0x8877u16), Ok(()));

b.flip_lo();

unsafe { assert_eq!(b.as_window_slice(), [ 0x78, 0x56, 0x34, 0x12
                                         , 0x44, 0x33, 0x22, 0x11
                                         , 0x77, 0x88 ]); }

unsafe fn unsafe_poke(&self, pos: u32, src: &[u8])

Writes the bytes at a given offset from the beginning of the window, into the supplied buffer. It is undefined behavior to write outside the iobuf window.

use iobuf::{RWIobuf,Iobuf};

let data = [ 1,2,3,4 ];

let mut b = RWIobuf::new(10);

unsafe {
  b.check_range_fail(1, 7);

  b.unsafe_advance(1);
  b.narrow();

  b.unsafe_poke(0, &data);
  b.unsafe_poke(3, &data);
  b.unsafe_advance(7);
}

b.flip_lo();

unsafe { assert_eq!(b.as_window_slice(), [ 1,2,3,1,2,3,4 ]); }

unsafe fn unsafe_poke_be<T: IntLike>(&self, pos: u32, t: T)

Writes a big-endian primitive at a given offset from the beginning of the window. It is undefined behavior to write outside the iobuf window.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::new(10);

unsafe {
  b.check_range_fail(0, 7);

  b.unsafe_poke_be(0, 0x0304u16);
  b.unsafe_poke_be(1, 0x0505u16);
  b.unsafe_poke_be(3, 0x06070809u32);
}

assert_eq!(b.resize(7), Ok(()));

unsafe { assert_eq!(b.as_window_slice(), [ 3, 5, 5, 6, 7, 8, 9 ]); }

unsafe fn unsafe_poke_le<T: IntLike>(&self, pos: u32, t: T)

Writes a little-endian primitive at a given offset from the beginning of the window. It is undefined behavior to write outside the iobuf window.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::new(10);

unsafe {
  b.check_range_fail(0, 7);

  b.unsafe_poke_le(0, 0x0304u16);
  b.unsafe_poke_le(1, 0x0505u16);
  b.unsafe_poke_le(3, 0x06070809u32);
}

assert_eq!(b.resize(7), Ok(()));

unsafe { assert_eq!(b.as_window_slice(), [ 4, 5, 5, 9, 8, 7, 6 ]); }

unsafe fn unsafe_fill(&mut self, src: &[u8])

Writes bytes from the supplied buffer, starting from the front of the window. It is undefined behavior to write outside the iobuf window.

After the bytes have been written, the window will be moved to no longer include then.

use iobuf::{RWIobuf,Iobuf};

let data = [ 1, 2, 3, 4 ];

let mut b = RWIobuf::new(10);

unsafe {
  b.check_range_fail(0, 8);

  b.unsafe_fill(&data[..]);
  b.unsafe_fill(&data[..]);
}

b.flip_lo();

unsafe { assert_eq!(b.as_window_slice(), [ 1,2,3,4,1,2,3,4 ]); }

unsafe fn unsafe_fill_be<T: IntLike>(&mut self, t: T)

Writes a big-endian primitive into the beginning of the window. It is undefined behavior to write outside the iobuf window.

After the primitive has been written, the window will be moved such that it is no longer included.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::new(10);

unsafe {
  b.check_range_fail(0, 10);

  b.unsafe_fill_be(0x12345678u32);
  b.unsafe_fill_be(0x11223344u32);
  // b.unsafe_fill_be(0x54321123u32); DO NOT DO THIS. Undefined behavior.
  b.unsafe_fill_be(0x8877u16);
}

b.flip_lo();

unsafe { assert_eq!(b.as_window_slice(), [ 0x12, 0x34, 0x56, 0x78
                                         , 0x11, 0x22, 0x33, 0x44
                                         , 0x88, 0x77 ]); }

unsafe fn unsafe_fill_le<T: IntLike>(&mut self, t: T)

Writes a little-endian primitive into the beginning of the window. It is undefined behavior to write outside the iobuf window.

After the primitive has been written, the window will be moved such that it is no longer included.

use iobuf::{RWIobuf,Iobuf};

let mut b = RWIobuf::new(10);

unsafe {
  b.check_range_fail(0, 10);

  b.unsafe_fill_le(0x12345678u32);
  b.unsafe_fill_le(0x11223344u32);
  // b.unsafe_fill_le(0x54321123u32); DO NOT DO THIS. Undefined behavior.
  b.unsafe_fill_le(0x8877u16);
}

b.flip_lo();

unsafe { assert_eq!(b.as_window_slice(), [ 0x78, 0x56, 0x34, 0x12
                                         , 0x44, 0x33, 0x22, 0x11
                                         , 0x77, 0x88 ]); }